Rename BUILD to BUILD.bazel
Best practices currently suggest calling the files BUILD.bazel and not
simply BUILD.

This is more explicit and creates less contention in projects that
have `build/` directories (which is quite common).  On case-insensitive
filesystems this can be particularly problematic.

Auto-generating tools (such as "Gazelle") are making files called
BUILD.bazel as well, so this is suggested for consistency in all non
legacy projects.

If both BUILD and BUILD.bazel files are present, the BUILD.bazel will
be preferred...further cementing its status as "canon".

https://github.com/bazelbuild/rules_go/issues/866
https://github.com/bazelbuild/bazel/issues/4517