Add a .bazelversion
It is suggested that Bazel be versioned by means of a .bazelversion
file, which contains a "semver" semantic version.

Tensorflow is using 3.7.2 at time of writing.  However, the current
version is 4.0.0, released Nov 2020 and the first "Long Term Support"
(LTS release):

https://blog.bazel.build/2020/11/10/bazel-4.0-announce.html

That makes it seem like a good version to build on, for being both
current and LTS.