Leverage that Google Test has BUILD.bazel
At the time the original tutorial was written, it seems that there was
no BUILD.bazel for the Google Gtest project.  That meant that this
project had to create its own.

This is no longer necessary:

https://stackoverflow.com/a/51847781

This commit removes the unnecessary `build_file` and updates the
README.md, as well as adding a couple of other comments regarding what
may not be intuitive to the Bazel uninitiated (e.g. me)